adjust skip params for surefire and forking action for failsafe
surefire and failsafe skip params are shared, there is no way to skip unit tests in same module/multimodule while specifying an IT to run from cmd line other than -Dtest=void. Therefore we
tell surefire not to fail if given -Dtest=sometestthatcannotbefound

Also failIfNoTests = false by default for surefire and failsafe which is a good thing for initial project setup and modules that do not need tests.

failsafe default should be to always fork tests - and so add a suitable 10 minute timeout to avoid potential hangs ( although developers are best advised to customize timeouts per test via a junit rule or something )